refactor: move common.ts in lib
