+++ src/devServer.ts
@@ -1,8 +1,8 @@
 import http from "node:http";
 
 import type { Config, Middleware } from "./config.js";
-import { pipe } from "./middleware/common.js";
-import type { Shared } from "./middleware/common.js";
+import { pipe } from "./middleware/lib/common.js";
+import type { Shared } from "./middleware/lib/common.js";
 
 export function startDevServer(config: Config = {}) {
   const shared: Shared = {};
