+++ src/middleware/staticFile.ts
@@ -3,7 +3,7 @@ import fs from "node:fs";
 
 import mime from "mime";
 
-import type { MiddlewareCreator } from "./common.js";
+import type { MiddlewareCreator } from "./lib/common.js";
 
 const staticFile: MiddlewareCreator = (config) => {
   const dir = path.resolve(config.prdServer?.dir || ".");
